Refactor SimpleContainer tests and improve code quality
Refactored SimpleContainerTests.cs into multiple files for better modularity and readability. Updated SimpleContainer.cs for performance and readability improvements. Added new test files for various aspects of SimpleContainer functionality. Created new service files to support tests. These changes enhance the structure, readability, and maintainability of the codebase.
