+++ src/Caliburn.Micro.Core.Tests/Services/Component.cs
@@ -1,0 +1,9 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    internal class Component : IComponent
+    {
+        public IDependency1 Dependency1 { get; set; }
+        public NonInterfaceDependency NonInterfaceDependency { get; set; }
+    }
+}
+
