+++ src/Caliburn.Micro.Core.Tests/Services/Dependency1.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Caliburn.Micro.Core.Tests.Services
+{
+    internal class Dependency1 : IDependency1
+    {
+        public IDependency2 Dependency2 { get; set; }
+        public IList<IEnumerableDependency> EnumerableDependencies { get; set; }
+    }
+}
+
