+++ src/Caliburn.Micro.Core.Tests/Services/Dependency2.cs
@@ -1,0 +1,4 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    internal class Dependency2 : IDependency2 { }
+}
