+++ src/Caliburn.Micro.Core.Tests/Services/EnumerableDependency1.cs
@@ -1,0 +1,8 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    internal class EnumerableDependency1 : IEnumerableDependency
+    {
+        public IDependency2 Dependency2 { get; set; }
+    }
+}
+
