+++ src/Caliburn.Micro.Core.Tests/Services/EnumerableDependency2.cs
@@ -1,0 +1,5 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    internal class EnumerableDependency2 : IEnumerableDependency { }
+
+}
