+++ src/Caliburn.Micro.Core.Tests/Services/IComponent.cs
@@ -1,0 +1,5 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    internal interface IComponent { }
+
+}
