+++ src/Caliburn.Micro.Core.Tests/Services/IDependency1.cs
@@ -1,0 +1,4 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    internal interface IDependency1 { }
+}
