+++ src/Caliburn.Micro.Core.Tests/Services/IDependency2.cs
@@ -1,0 +1,4 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    internal interface IDependency2 { }
+}
