+++ src/Caliburn.Micro.Core.Tests/Services/IEnumerableDependency.cs
@@ -1,0 +1,4 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    internal interface IEnumerableDependency { }
+}
