+++ src/Caliburn.Micro.Core.Tests/Services/NonInterfaceDependency.cs
@@ -1,0 +1,4 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    internal class NonInterfaceDependency { }
+}
