+++ src/Caliburn.Micro.Core.Tests/Services/SecondDependency1.cs
@@ -1,0 +1,4 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    internal class SecondDependency1 : Dependency1 { }
+}
