+++ src/Caliburn.Micro.Core.Tests/Services/SingleEmptyConstructorType.cs
@@ -1,0 +1,10 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    public class SingleEmptyConstructorType
+    {
+        public SingleEmptyConstructorType()
+        {
+
+        }
+    }
+}
