+++ src/Caliburn.Micro.Core.Tests/Services/SingleIntConstructor.cs
@@ -1,0 +1,13 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    public class SingleIntConstructor
+    {
+        public int Value { get; private set; }
+
+        public SingleIntConstructor(int x)
+        {
+            this.Value = x;
+        }
+    }
+
+}
