+++ src/Caliburn.Micro.Core.Tests/Services/SingleNonEmptyConstructorType.cs
@@ -1,0 +1,9 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    public class SingleNonEmptyConstructorType
+    {
+        public SingleNonEmptyConstructorType(SingleEmptyConstructorType type)
+        {
+        }
+    }
+}
