+++ src/Caliburn.Micro.Core.Tests/Services/TestService.cs
@@ -1,0 +1,5 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    internal class TestService : ITestService { }
+    internal class AnotherTestService : ITestService { }
+}
