+++ src/Caliburn.Micro.Core.Tests/Services/TestServiceWithDependency.cs
@@ -1,0 +1,7 @@
+﻿namespace Caliburn.Micro.Core.Tests.Services
+{
+    internal class TestServiceWithDependency
+    {
+        public ITestService Dependency { get; set; }
+    }
+}
